Add Collada skeletal animation support and refactor renderer into partial classes
- Implement matrix-based animation export for Collada (.dae) files
- Animation data from .dba files now exports with proper bone channel targeting
- Split ColladaModelRenderer (1691 lines) into 7 partial class files:
  - Animation.cs: Animation export with keyframe sampling
  - Materials.cs: Material/texture creation
  - Geometry.cs: Mesh processing
  - Skeleton.cs: Controller/bone/skinning
  - Nodes.cs: Visual scene hierarchy
  - Utilities.cs: String formatting helpers
- Fix bone name consistency for animation channel targeting
